page5->page2 화면 전환 수정 중

+++ app/src/main/java/com/example/brandol/searchCategory/UserCategoryFragment.kt
@@ -108,7 +108,7 @@ class UserCategoryFragment : Fragment() {
                 navigateToOpponentAvatar(userId)
             }
 
-            override fun onItemClick(userId: Int) {
+            override fun onItemClick(userId: Long) {
                 TODO("Not yet implemented")
             }
         }
